Fix some "implicit conversion loses integer precision" warnings (#3375)
Motivation:
Build SwiftPM with `-Xcc -Wshorten-64-to-32` and see a bunch of these warnings.

Modifications:
Fix warnings wherever we can. Many of them come from the dependencies though.

+++ Sources/PackageCollectionsSigningLibc/ocsp_cl.c
@@ -55,7 +55,7 @@ OCSP_ONEREQ *OCSP_request_add0_id(OCSP_REQUEST *req, OCSP_CERTID *cid)
 
 int OCSP_response_status(OCSP_RESPONSE *resp)
 {
-    return ASN1_ENUMERATED_get(resp->responseStatus);
+    return (int)ASN1_ENUMERATED_get(resp->responseStatus);
 }
 
 /*
